homework_3/requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
scikit-learn>=1.0.2
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4 